multibandDriver: write metadata from detection
We now put some useful information in the detection metadata, so we need
to make sure it's clean and write it out.

Not using MultiBandDriverTask.writeMetadata because the 'dataRef' that's
provided there is actually a list of data references, and it's not clear
we can identify which entry in the metadata belongs to which dataRef.