Fix running detections
Detections were not being run when none were present. This lead to
the detection stage never being run, but if detections were available
somewhere above the direct parent this task was still treating them
as run. Change this behavior so that if no detections are the task
aborts or if doDetection is set detection will be forced to run.