Add FieldTractDataIdContainer to multiBandDriver
We use both field and tract, even though they are redundant.
The field is useful to the user, while tract is required by the system.
But this means that DISTINCT on field, filter gets all of the tracts
with no WHERE on field.
and the tract is just added separately, it's not part of the query
(because the DB doesn't know about tracts).