Add Gen3 implementation to skyCorrection.py.
The new Gen3 implementation of the skyCorrection code is serialized,
and duplicates some of the SkyCorrectionTask methods to provide this
serialized form via a run() method (which matches the Gen2
runDataRef() method).  Comments have been added to link the Gen2 and
Gen3 methods, and both methods should be updated for any future code
changes.

An additional "rawLinker" input dataset is needed for Gen3, as the
calExp (which has the dataset.visit column populated) cannot find a
link to find the appropriate camera/sky calibration entries.  Adding
this rawLinker allows the calExp to connect to the raw data that was
its source, and use that connection to find the calibration files.
This dataset is used only for constructing the quantum graph, and is
removed prior to processing in the task run() method.