Make use of Gen3-mimicking getGen3CoaddExposureId
This is temporary functionality to aid with the Gen2 to Gen3 middleware
migration.  In particular, the ID integer associated with a particular
coadd exposure was computed differently between the two middlewares.
This number is used in places a seed for random number generation used
in certain algorithms (e.g. dynamic detection).  This seed value must
match in both middlewares if parity is to be achieved, so here we update
the expId evaluation in Gen2 to match that of Gen3.